turnip: use 32-bit offset in tu_cs_entry
We don't support nor expect BOs to be that big in tu_cs.

+++ src/freedreno/vulkan/tu_private.h
@@ -484,7 +484,7 @@ struct tu_cs_entry
    const struct tu_bo *bo;
 
    uint32_t size;
-   uint64_t offset;
+   uint32_t offset;
 };
 
 enum tu_cs_mode
